% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2pptx.R
\name{pdf2pptx}
\alias{pdf2pptx}
\title{Convert PDF to PPTX}
\usage{
pdf2pptx(pdf_filename, pptx_filename, dpi = 300)
}
\arguments{
\item{pdf_filename}{File name of the source PDF file}

\item{pptx_filename}{File name of the destination PPTX file}

\item{dpi}{Optional parameter specifying image density for rendering images from PDF}
}
\value{
Nothing
}
\description{
Convert PDF to PPTX
}
\examples{
example_from_url <-
  "http://mirrors.ctan.org/macros/latex/contrib/beamer-contrib/themes/metropolis/demo/demo.pdf"
# conversion takes several seconds
pdf2pptx(example_from_url, "demo.pptx")
unlink("demo.pptx")
}
